Move castSenderApi to plugin dir
